pyproject.toml
[project]
name = "mybot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "bidict>=0.23.1",
    "dishka>=1.7.2",
    "faiss-cpu>=1.13.1",
    "fastapi>=0.127.0",
    "google-genai>=1.55.0",
    "loguru>=0.7.3",
    "numpy>=2.4.0",
    "openai>=2.11.0",
    "pydantic-settings>=2.12.0",
    "uvicorn[standard]>=0.40.0",
]
